Further developed the sprint function
I developed the sprint function further by adding a progress bar below the pc to show how much of it is left, problem with it is that it is not entirely accurate. I added an over sprint function to it so that when you hit a certain thresh hold, you have to wait till you can sprint again, problem with it is that the progress bar is not accurate and so you have to wait for a while after it is full to be able to sprint again. The stamina bar is also very big and not the same pattern as the health bar so I will need help fixing that.